Using a new awsbox which supports local scripts to copy sauce credentials to ephemeral instances and start sauce tests.
