+++ lockdown.json
@@ -21,7 +21,7 @@
     "0.0.5": "971b8995078d83c80f2372f134c496e71b293a46"
   },
   "awsbox": {
-    "0.3.1": "bbd433489d7d73c8b38c4f60f39ab6b034e73584"
+    "0.3.2": "*"
   },
   "bcrypt": {
     "0.7.1": "923e2623331211adcab6ac84ec4fcd41713e1e58"
