+++ package.json
@@ -1,59 +1,60 @@
 {
-    "name": "browserid",
-    "version": "1.0.0-b1",
-    "private": true,
-    "dependencies": {
-        "JSONSelect": "0.4.0",
-        "bcrypt": "0.7.1",
-        "compute-cluster": "0.0.6",
-        "connect": "1.7.2",
-        "convict": "0.0.6",
-        "cjson": "0.0.6",
-        "client-sessions": "0.0.9",
-        "connect-cachify": "0.0.11",
-        "connect-cookie-session": "0.0.2",
-        "connect-logger-statsd": "0.0.1",
-        "ejs": "0.8.3",
-        "etagify": "0.0.2",
-        "express": "2.5.0",
-        "gobbledygook": "0.0.3",
-        "lockdown": "0.0.3",
-        "jwcrypto": "0.4.2",
-        "mysql": "0.9.5",
-        "nodemailer": "0.3.21",
-        "mkdirp": "0.3.0",
-        "optimist": "0.2.8",
-        "postprocess": "0.2.4",
-        "semver": "1.0.12",
-        "temp": "0.4.0",
-        "uglify-js": "1.0.6",
-        "uglifycss": "0.0.5",
-        "underscore": "1.3.1",
-        "urlparse": "0.0.1",
-        "validator": "0.4.9",
-        "winston": "0.6.2"
-    },
-    "optionalDependencies": {
-        "node-statsd": "https://github.com/downloads/lloyd/node-statsd/0509f85.tgz"
-    },
-    "devDependencies": {
-        "vows": "0.5.13",
-        "awsbox": "0.3.1",
-        "irc": "0.3.3",
-        "jshint": "0.7.1",
-        "minimatch": "0.2.6",
-        "which": "1.0.5",
-        "htmlparser": "1.7.6",
-        "node-inspector": "0.2.0beta3"
-    },
-    "scripts": {
-        "preinstall": "node ./scripts/lockdown",
-        "postinstall": "node ./scripts/postinstall.js",
-        "postupdate": "node ./scripts/postinstall.js",
-        "test": "./scripts/test",
-        "start": "node ./scripts/run_locally.js"
-    },
-    "engines": {
-        "node": ">= 0.8.11"
-    }
+  "name": "browserid",
+  "version": "1.0.0-b1",
+  "private": true,
+  "dependencies": {
+    "JSONSelect": "0.4.0",
+    "bcrypt": "0.7.1",
+    "compute-cluster": "0.0.6",
+    "connect": "1.7.2",
+    "convict": "0.0.6",
+    "cjson": "0.0.6",
+    "client-sessions": "0.0.9",
+    "connect-cachify": "0.0.11",
+    "connect-cookie-session": "0.0.2",
+    "connect-logger-statsd": "0.0.1",
+    "ejs": "0.8.3",
+    "etagify": "0.0.2",
+    "express": "2.5.0",
+    "gobbledygook": "0.0.3",
+    "lockdown": "0.0.3",
+    "jwcrypto": "0.4.2",
+    "mysql": "0.9.5",
+    "nodemailer": "0.3.21",
+    "mkdirp": "0.3.0",
+    "optimist": "0.2.8",
+    "postprocess": "0.2.4",
+    "semver": "1.0.12",
+    "temp": "0.4.0",
+    "uglify-js": "1.0.6",
+    "uglifycss": "0.0.5",
+    "underscore": "1.3.1",
+    "urlparse": "0.0.1",
+    "validator": "0.4.9",
+    "winston": "0.6.2",
+    "awsbox": "git://github.com/shane-tomlinson/awsbox.git#9b456826f77d4e200468f287558af9a4c9935293"
+  },
+  "optionalDependencies": {
+    "node-statsd": "https://github.com/downloads/lloyd/node-statsd/0509f85.tgz"
+  },
+  "devDependencies": {
+    "vows": "0.5.13",
+    "awsbox": "git://github.com/shane-tomlinson/awsbox.git#9b456826f77d4e200468f287558af9a4c9935293",
+    "irc": "0.3.3",
+    "jshint": "0.7.1",
+    "minimatch": "0.2.6",
+    "which": "1.0.5",
+    "htmlparser": "1.7.6",
+    "node-inspector": "0.2.0beta3"
+  },
+  "scripts": {
+    "preinstall": "node ./scripts/lockdown",
+    "postinstall": "node ./scripts/postinstall.js",
+    "postupdate": "node ./scripts/postinstall.js",
+    "test": "./scripts/test",
+    "start": "node ./scripts/run_locally.js"
+  },
+  "engines": {
+    "node": ">= 0.8.11"
+  }
 }
