--- scripts/awsbox/post_start.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-scripts/show_config.js
-
-
-
