+++ scripts/awsbox_remote/post_create.sh
