+++ scripts/awsbox_remote/post_start.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+node automation-tests/scripts/post-update.js
+
+scripts/show_config.js
